requirements.txt
PyYAML>=6.0.1
typer>=0.9.0
PyInquirer>=1.0.3
rich>=13.5.2
pytest>=7.2.2
requests>=2.31.0
python-dotenv~=0.21.1
path~=16.7.1
selenium>=4.8.0
behave>=1.2.6